Made newPostPage functional
