+++ src/components/redditClone/index.js
@@ -2,7 +2,6 @@ import React from 'react'
 import _ from 'lodash'
 import Header from 'components/redditClone/header.js'
 import Posts from 'components/redditClone/posts.js'
-import { newPostPage } from 'redux/actions/posts'
 
 export default props => {
   return props.loading === true
@@ -10,6 +9,6 @@ export default props => {
     : <div>
         <Header />
         {_.map(props.posts, post => <Posts post={post} />)}
-        <button onClick={newPostPage}>Add a New Post</button>
+        <a href="/newPost">Add a New Post</a>
       </div>
 }
