+++ src/components/redditClone/newPosts/newPost.js
